# Optimize multiplication