1. Fully unroll `umul()` implementation (the first commit does majority of that, the second is small tuning giving small improvement, but affects readability).
2. The use the `umul()` implementation to implement `Mul()`.
3. Also optimize `Squared()`.

```
name                      old time/op  new time/op  delta
_Mul/single/big-6          210ns ± 1%   210ns ± 0%     ~     (p=1.000 n=10+10)
_Mul/single/uint256-6     36.3ns ± 0%  18.0ns ± 0%  -50.37%  (p=0.000 n=10+10)
_Square/single/big-6       207ns ± 0%   207ns ± 1%     ~     (p=0.108 n=9+10)
_Square/single/uint256-6  34.9ns ± 0%  13.7ns ± 0%  -60.71%  (p=0.000 n=9+9)
_Exp/large/big-6          47.5µs ± 1%  47.6µs ± 1%     ~     (p=0.393 n=10+10)
_Exp/large/uint256-6      14.4µs ± 0%   6.8µs ± 0%  -52.91%  (p=0.000 n=10+9)
_Exp/small/big-6          14.2µs ± 0%  14.2µs ± 1%     ~     (p=0.218 n=10+10)
_Exp/small/uint256-6      1.23µs ± 0%  0.57µs ± 0%  -53.43%  (p=0.000 n=10+9)
MulMod/small/uint256-6    85.9ns ± 0%  75.8ns ± 0%  -11.71%  (p=0.000 n=8+10)
MulMod/small/big-6         135ns ± 0%   135ns ± 0%     ~     (all equal)
MulMod/mod64/uint256-6     228ns ± 1%   216ns ± 0%   -5.21%  (p=0.000 n=10+10)
MulMod/mod64/big-6         594ns ± 0%   594ns ± 0%     ~     (p=0.596 n=9+9)
MulMod/mod128/uint256-6    422ns ± 0%   391ns ± 0%   -7.43%  (p=0.000 n=10+8)
MulMod/mod128/big-6       1.11µs ± 0%  1.11µs ± 0%     ~     (p=0.366 n=10+10)
MulMod/mod192/uint256-6    422ns ± 0%   386ns ± 0%   -8.64%  (p=0.000 n=10+8)
MulMod/mod192/big-6       1.02µs ± 0%  1.02µs ± 0%     ~     (p=0.526 n=9+10)
MulMod/mod256/uint256-6    419ns ± 0%   376ns ± 0%  -10.47%  (p=0.000 n=10+10)
MulMod/mod256/big-6        930ns ± 0%   927ns ± 0%   -0.32%  (p=0.000 n=10+10)
```