@chfast 2020-04-14
> Generally looks good, but I haven't checked how good test coverage we have of `Squared`. Is it properly covered?

Not sure if properly tested (although I was experiencing failures during development), but it has _some_ coverage: https://codecov.io/gh/holiman/uint256/src/9bfa06571ed6d28a42676b96307e127875a3c709/uint256.go#L292